requirements.txt
alembic==1.11.1
anyio==3.7.0
autopep8==2.0.2
bcrypt==4.0.1
black==23.7.0
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.98.0
filelock==3.12.2
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
identify==2.5.24
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
orjson==3.9.1
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.8.1
pluggy==1.2.0
pre-commit==3.3.3
psycopg2==2.9.6
pyasn1==0.5.0
pycodestyle==2.10.0
pydantic==1.10.9
pyflakes==3.0.1
pytest==7.4.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.23
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.3
ujson==5.8.0
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.23.1
watchfiles==0.19.0
websockets==11.0.3
